Much easier, leaner core lib [untested, breaking]
New user-side interface is much like before:

  local shine = require 'shine'

  filmgrain = shine.filmgrain()
  filmgrain.grainsize = 2

  effect = shine(filmgrain).next(shine.crt()).next(shine.desaturate())
  effect.strength = 0.8 -- sets property of desaturate filter
  effect.opacity = 0.3 -- properties affect the first match from the end
                       -- of the chain (here: filmgrain)

  function love.draw()
    effect:draw(function() ... end)
  end

Shader-Developer interface is much simpler than before: Each effect is a
function that must return a table with the key "shader" that points to a
shader or "draw" that points to a function(buffer, shader).  Here,
buffer is a function provided by shine that provides a front and back
buffer and shader is the table entry "shader" (may be nil).  The keys
"setters" and "defaults" may contain setters for shader parameters (like
"grainsize" above) and the corresponding default values.

For usage examples, see the next commits.